+++ src/main/java/com/cmpscjg/profiles/utils/DataHelper.java
@@ -1,0 +1,12 @@
+package com.cmpscjg.profiles.utils;
+
+import com.cmpscjg.profiles.Profiles;
+
+public class DataHelper {
+
+    private Profiles plugin;
+
+    public DataHelper(Profiles plugin) {
+        this.plugin = plugin;
+    }
+}